log-processor/requirements.txt
"""
Dependencies for log processor service.

Key packages:
- kafka-python: Kafka consumer client
- elasticsearch: Elasticsearch Python client
- psycopg2: PostgreSQL adapter
- scikit-learn: Machine learning library (Isolation Forest)
"""

kafka-python==2.0.2
elasticsearch==8.10.0
psycopg2-binary==2.9.9
scikit-learn==1.3.2
numpy==1.24.3
